\name{data.example}
\docType{data}
\alias{data.example}
\title{Transaction Data of The Product}
\description{
  This data set is an example transaction data set
  which contains transactional details of a product in a couple of years
}
\usage{data.example}
\format{A data.frame containing 975 observations.}
\keyword{datasets}
