% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigerhitteR.R
\name{dateRefill.fromData}
\alias{dateRefill.fromData}
\title{Complete the hollow dataset}
\usage{
dateRefill.fromData(data, dateCol, fixedVec, pChanged, pChangedNum)
}
\arguments{
\item{data}{The data.frame dataset which is ready to be processed}

\item{dateCol}{Date column}

\item{fixedVec}{A row of column number which should be kept same values with the original}

\item{pChanged}{The column number which should be changed to different value into new record.}

\item{pChangedNum}{The value of a specific column which should be put into the new record.}
}
\value{
The dataset which is completed.
}
\description{
Take time series dataset and fields, then refill the missing date records and other fields.
}
\details{
Real time series sales dataset could be not continuous in 'date' field. e.g., monthly sales data is continuous,
 but discrete in daily data.

 This hollow dataset is not complete for time series analysis. Function dateRefill.fromFile
 is a transformation which tranforms uncomplete dataset into complete dataset.
}
\examples{
# mydata <- data.example
# mydata.final <- dateRefill.fromData(data = mydata,dateCol = 2,fixedVec = c(3:10),
#                                     pChanged = 11,pChangedNum = 0)
}
\author{
Will Kuan
}

