% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigerhitteR.R
\name{dateRefill.fromsFile}
\alias{dateRefill.fromsFile}
\title{Complete the hollow dataset}
\usage{
dateRefill.fromsFile(inPath, sheet, dateCol, outPath, fixedVec, pChanged,
  pChangedNum)
}
\arguments{
\item{inPath}{A path which is the location of uncompleted dataset which must be xlsx file}

\item{sheet}{A worksheet name of the dataset}

\item{dateCol}{Date column}

\item{outPath}{A path where the location of xlsx file of completed dataset should be}

\item{fixedVec}{A row of column number which should be kept same values with the original}

\item{pChanged}{The column number which should be changed to different value into new record.}

\item{pChangedNum}{The value of a specific column which should be put into the new record.}
}
\description{
Take time series dataset and fields, then refill the missing date records and other fields.
}
\details{
Real time series sales dataset could be not continuous in 'date' field. e.g., monthly sales data is continuous,
 but discrete in daily data.

 This hollow dataset is not complete for time series analysis. Function dateRefill.fromFile
 is a transformation which tranforms uncomplete dataset into complete dataset.
}
\examples{
# Please refer to the examples of function dateRefill.fromData
}
\author{
Will Kuan
}

